Complete FastCollinearPoints (tests pasts).

+++ week3_twoClassicSortingAlgorithms/assignment/Point.java
@@ -75,7 +75,7 @@ public class Point implements Comparable<Point> {
         } else if (this.x == that.x) {
             return Double.POSITIVE_INFINITY;
         } else {
-            return (double)(that.y - this.y) / (that.x - this.x);
+            return (double)(that.y - this.y) / (double) (that.x - this.x);
         }
     }
